clean up extra files and comments, reorganize file locations, added width height props for scalable display
