+++ src/pages/Favorites/Favorites.tsx
@@ -6,9 +6,7 @@ import { selectFavoritedResources } from '../../store/resourcesSlice';
 import { withStyles } from '@material-ui/core/styles';
 import { styles } from './FavoriteStyles';
 import { Link } from 'react-router-dom';
-import { FormControl, Select, MenuItem, Input } from '@material-ui/core';
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
-import FilterDropdown from './FilterDropdown';
+import FilterDropdown from '../../components/Filters/FilterDropdown';
 
 function Favorites({ match, classes }) {
   const favResources = useSelector(selectFavoritedResources);
@@ -18,22 +16,8 @@ function Favorites({ match, classes }) {
   );
   const allTypes : string[] = ['Articles', 'Videos']
   
-
-  /* begin joyce task */
-
-  // set up usestate to alter a list of topics we will filter by
   const [currTopics, setTopic] = useState<Set<string>>(new Set(allTopics));
   const [resType, setresType] = useState<Set<string>>(new Set(allTypes));
-
-
-  /* end joyce task */
-
-  // const handleResChange = (event) => {
-  //   setresType(event.target.value);
-  // };
-  // const handleTopicChange = (event) => {
-  //   setTopic(event.target.value);
-  // };
 
   function filteredFavResources(resource: any) {
     if (
@@ -60,8 +44,7 @@ function Favorites({ match, classes }) {
       return <span />;
     }
   }
-
-  // 
+ 
   return (
     <div className={classes.page}>
       <div className={classes.header}>
@@ -78,46 +61,6 @@ function Favorites({ match, classes }) {
           </FilterDropdown>
         </div>
       </div>
-      
-      
-      {/* <div className={classes.filters}>
-        <div className={classes.filterLabel}>Topics</div>
-        <FormControl className={classes.formControl}>
-            <Select
-              multiple
-              value={currTopics}
-              onChange={handleTopicChange}
-              // input={<Input />}
-              disableUnderline
-              displayEmpty
-              // renderValue={value => 'All'}
-              IconComponent={() => (<ExpandMoreIcon />)}
-              className={classes.select}
-              classes={{ selectMenu: classes.selectMenu, icon: classes.icon }}
-              MenuProps={{ classes: { paper: classes.dropdownStyle }}}
-            >
-              {Object.keys(topics).map((key) => (
-                <MenuItem key={key} value={topics[key].name}>
-                  {topics[key].name}
-                </MenuItem>
-              ))}
-            </Select>
-        </FormControl>
-        <div className={classes.filterLabel}>Types</div>
-        <FormControl className={classes.formControl}>
-          <Select
-            value={resType}
-            disableUnderline
-            className={classes.select}
-            classes={{ selectMenu: classes.selectMenu, icon: classes.icon }}
-            onChange={handleResChange}
-          >
-            <MenuItem value={'All'}>All</MenuItem>
-            <MenuItem value={'Video'}>Videos</MenuItem>
-            <MenuItem value={'Article'}>Articles</MenuItem>
-          </Select>
-        </FormControl>
-      </div> */}
       {Object.keys(favResources).map(filteredFavResources)}
     </div>
   )};
